jpegxl/interop-2026/features/vardct-small-blocks.html

<!DOCTYPE html>
<title>JPEG XL interop: VarDCT small block stress</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes tiny noisy VarDCT fixture.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/interop-helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('../resources/8x8_noise.jxl');
  assert_equals(img.naturalWidth, 8);
  assert_equals(img.naturalHeight, 8);
}, 'Small VarDCT-oriented fixture decodes.');

</script>
